move supplier API into its own file
- add a negative test where we fail to create a supplier
- add the ability to delete a supplier by id (added a test too)
